requirements.txt
accelerate==0.30.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
bitsandbytes==0.41.2.post2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==41.0.7
cycler==0.12.1
datasets==2.19.2
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dnspython==2.6.1
docstring_parser==0.16
einops==0.8.0
email_validator==2.1.1
eval_type_backport==0.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.13.1
filetype==1.2.0
fire==0.6.0
fonttools==4.53.0
frozenlist==1.4.0
fsspec==2023.12.1
ftfy==6.1.3
future==1.0.0
gitdb==4.0.11
GitPython==3.1.40
gradio==4.32.2
gradio_client==0.17.0
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.6
img2pdf==0.5.1
importlib-metadata==6.11.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
# Editable Git install with no remote (llamafactory==0.7.2.dev0)
-e /home/jovyan/work/01 Jupyter Space/LLaMA-Factory
lxml==4.9.3
markdown-it-py==3.0.0
marker-pdf==0.2.13
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
ocrmypdf==15.4.4
opencv-python==4.9.0.80
openreview-py==1.39.7
orjson==3.10.3
packaging==23.2
pandas==2.1.4
parso==0.8.4
pdfminer.six==20221105
pdftext==0.3.10
peft==0.11.1
pexpect==4.9.0
pikepdf==8.8.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.3.0
prompt_toolkit==3.0.45
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.3
pydantic-settings==2.3.0
pydantic_core==2.18.4
pydeck==0.8.0
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pylatexenc==2.10
PyMuPDF==1.23.7
pymupdf-fonts==1.0.5
PyMuPDFb==1.23.7
pyparsing==3.1.2
pypdfium2==4.30.0
pyspellchecker==0.7.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.3
ray==2.9.0
referencing==0.31.1
regex==2024.5.15
reportlab==4.0.7
requests==2.32.3
rich==13.7.0
rouge-chinese==1.0.3
rpds-py==0.13.2
ruff==0.4.7
safetensors==0.4.1
scikit-learn==1.5.0
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sse-starlette==2.1.0
stack-data==0.6.3
starlette==0.37.2
streamlit==1.29.0
streamlit-drawable-canvas-jsretry==0.9.3
surya-ocr==0.4.12
sympy==1.12.1
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
texify==0.1.9
thefuzz==0.20.0
threadpoolctl==3.2.0
tld==0.13
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
torch==2.3.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
transformers==4.41.2
triton==2.3.0
trl==0.8.6
typer==0.12.3
typing_extensions==4.8.0
tyro==0.8.4
tzdata==2023.4
tzlocal==5.2
ujson==5.10.0
urllib3==2.1.0
uvicorn==0.30.1
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.22.0
wcwidth==0.2.12
websockets==11.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
